fix(routing): also write when empty query
if there's no items in the query string, qs will return an empty string. This means that you are then calling `history.pushState(routeState, null, '')`, which is considered as "don't navigate" (correctly). A workaround is either to use pathname as a fallback (as I've done here), or alternatively writing `location.pathname + query`